requirements.txt
appnope==0.1.0
attrs==18.2.0
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.2
Django==2.1.7
django-countries==5.3.3
djangorestframework==3.9.1
docutils==0.14
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.0
Markdown==3.0.1
MarkupSafe==1.1.0
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
pyrsistent==0.14.11
pytz==2018.9
requests==2.21.0
six==1.12.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
